14_Agent/crew_ai_YoutubeChannelSearchTool/requirements.txt
crewai
crewai_tools
load_dotenv
langchain
langchain-openai
langchain-google-genai==1.0.4
langchain-huggingface
python-dotenv
